Move to a new widgets/dock.py the code for SpyderDockWidget
There is no need to have this in api/plugins.py because it's not
supposed to be touched by plugin creators.
